+++ test/foo/service.clj
@@ -17,7 +17,8 @@
 
 
 (defn ^{:inject :svc/create-order} service-create-order
-  [^:inject find-items ^:inject db-create-order user-details order-details]
+  [^:inject [find-items db-create-order]
+   user-details order-details]
   (let [item-ids   (find-items (:item-ids order-details))
         order-data {:order-data :dummy}]  ; prepare order-data
     (db-create-order order-data)))