+++ test/foo/web.clj
@@ -16,7 +16,9 @@
 
 
 (defn web-create-order
-  [^:inject find-user ^{:inject :svc/create-order} service-create-order web-request]
+  [^:inject {:keys [find-user]
+             service-create-order :svc/create-order
+             :as all} web-request]
   (let [user-details (find-user (:session web-request))
         order-details {:order :dummy-order}
         created-order (service-create-order user-details order-details)]