+++ test/foo/service.clj
@@ -7,13 +7,29 @@
 ;   You must not remove this notice, or any other, from this software.
 
 
-(ns foo.service)
+(ns foo.service
+  (:require
+    [dime.core :as di]))
 
 
 (defn find-items
   "This has an implicit inject name, hence will be overridden by the explicit one in foo.db namespace."
   [item-ids]
   :mock-items)
+
+
+(defn ^:inject recommend-products
+  "Return item IDs for specified user ID."
+  [^:inject items-cache user-id]
+  (let [item-ids (get items-cache user-id)]
+    (find-items item-ids)))
+
+
+(di/definj service-browse-items
+  [find-items recommend-products]
+  [user-id]
+  (let [item-ids (recommend-products user-id)]
+    (find-items item-ids)))
 
 
 (defn ^{:inject :svc/create-order} service-create-order