+++ test/dime/core_test.clj
@@ -29,7 +29,7 @@
       (is (= (di-p :second :third) {:data [:first :second :third]})))))
 
 
-(def ten-times (di/inj {:inject :ten-multiplier} [^{:inject :multiplier} n] (* 10 ^long n)))
+(def ten-times (di/inj {:expose :ten-multiplier} [^{:inject :multiplier} n] (* 10 ^long n)))
 
 
 (deftest test-inj
@@ -39,7 +39,7 @@
 
 (deftest test-assoc-inj
   (is (contains? (di/assoc-inj {} ten-times) :ten-multiplier))
-  (is (contains? (di/assoc-inj {} ten-times (di/inj {:inject :foo-bar} [foo bar] [foo bar])) :foo-bar)))
+  (is (contains? (di/assoc-inj {} ten-times (di/inj {:expose :foo-bar} [foo bar] [foo bar])) :foo-bar)))
 
 
 (defn sample-0