+++ CHANGES.md
@@ -26,7 +26,7 @@
   * [TODO] Support for creating injectable alias to pre-existing ordinary vars
   * [TODO] Support for updating pre-existing var metadata with injection info
 * Non-var injectables
-  * [TODO] Make `dime.core/inj` read metadata using attribute names in `dime.util`
+  * Make `dime.core/inj` read metadata using attribute names in `dime.util`
   * [TODO] Custom, efficient injectable types as defrecord
   * [TODO] Discoverable as a var